Fixed compilation errors on Win32 platform (refactored VkDestroyer class).

+++ Library/Source Files/04 Resources and Memory/11 Mapping, updating and unmapping host-visible memory.cpp
@@ -45,7 +45,7 @@ namespace VulkanCookbook {
       return false;
     }
 
-    std::memcpy( local_pointer, data, data_size );
+    std::memcpy( local_pointer, data, static_cast<size_t>(data_size) );
 
     std::vector<VkMappedMemoryRange> memory_ranges = {
       {
